Refactoring: Merge classes FeedItemTest and FeedItemTest2
